Allow reducing inside goals with hyphen notation
If you add a hyphen after a goal, as in:

    ?help-

Then it will display it with paths, as in:

    Goal ?help:
    With type: Equal-000(Nat-100)(id-010(zero-110))(zero-1)

If you then add the path after the goal, it will reduce that path, i.e.,

    ?help-110

Will show the goal as:

    Goal ?aaa:
    With type: Equal(Nat)(id((P) (z) (s) z))(zero)

That is, the `zero-110` reference was reduced. Being able to reduce
specific parts of the goal will make theorem proving easier.